+++ src/main/java/com/robocraft999/amazingtrading/registry/ATBlocks.java
@@ -1,7 +1,7 @@
 package com.robocraft999.amazingtrading.registry;
 
-import com.robocraft999.amazingtrading.content.shredder.CreateShredderBlock;
 import com.robocraft999.amazingtrading.content.shop.ShopBlock;
+import com.robocraft999.amazingtrading.content.shredder.CreateShredderBlock;
 import com.simibubi.create.content.kinetics.BlockStressDefaults;
 import com.simibubi.create.foundation.data.AssetLookup;
 import com.simibubi.create.foundation.data.SharedProperties;
@@ -11,6 +11,7 @@ import net.minecraft.world.level.material.MapColor;
 
 import static com.robocraft999.amazingtrading.AmazingTrading.REGISTRATE;
 import static com.simibubi.create.foundation.data.ModelGen.customItemModel;
+import static com.simibubi.create.foundation.data.TagGen.axeOnly;
 import static com.simibubi.create.foundation.data.TagGen.pickaxeOnly;
 
 public class ATBlocks {
@@ -29,6 +30,7 @@ public class ATBlocks {
             .block("shop", ShopBlock::new)
             .initialProperties(SharedProperties::stone)
             .properties(p -> p.noOcclusion().mapColor(MapColor.METAL))
+            .transform(axeOnly())
             .blockstate((c, p) -> p.horizontalBlock(c.getEntry(), AssetLookup.partialBaseModel(c, p), 90))
             .transform(BlockStressDefaults.setImpact(16))
             .addLayer(() -> RenderType::cutoutMipped)